Separering av klasser, test
